requirements.txt
chardet==3.0.4
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
dnspython==1.16.0
Flask==1.1.2
Flask-Cors==3.0.9
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
matplotlib
numpy==1.19.5
pandas==1.1.5
pyparsing==2.4.7
requests==2.24.0
urllib3==1.25.11
cassandra-driver==3.25.0
astrapy==0.0.2
simplejson==3.17.2
